+++ src/__tests__/ignore-file.test.ts
@@ -12,7 +12,7 @@ test('ignorefile synthesises correctly', () => {
   expect(splitAndIgnoreMarker(synthSnapshot(prj)['.dockerignore'])).toEqual([]);
 });
 
-test('ignorefile does not sort entries', () => {
+test('ignorefile puts includes (!) after excludes and sorts the entries', () => {
   // GIVEN
   const prj = new TestProject();
 
@@ -24,12 +24,12 @@ test('ignorefile does not sort entries', () => {
 
   // THEN
   expect(splitAndIgnoreMarker(synthSnapshot(prj)['.dockerignore'])).toEqual([
-    '!c.txt',
-    '!d.txt',
     'a.txt',
     'b.txt',
     'e.txt',
     'f.txt',
+    '!c.txt',
+    '!d.txt',
   ]);
 });
 
@@ -59,7 +59,7 @@ test('ignorefile excludes file after inclusion and exclusion', () => {
 
   // THEN
   expect(splitAndIgnoreMarker(synthSnapshot(prj)['.dockerignore'])).toEqual([
-    'a.txt',
+    '!a.txt',
   ]);
 });
 
@@ -80,6 +80,84 @@ test('ignorefile omits duplicated includes and excludes', () => {
     'b.txt',
     '!c.txt',
     '!d.txt',
+  ]);
+});
+
+test('includes (!) are always at the end', () => {
+  // GIVEN
+  const prj = new TestProject();
+  const ignore = new IgnoreFile(prj, '.myignorefile');
+
+  // WHEN
+  ignore.addPatterns('*.foo');
+  ignore.addPatterns('!hello.foo');
+  ignore.include('negated.txt');
+  ignore.addPatterns('*.bar');
+  ignore.exclude('*.bar');
+  ignore.include('zoo.foo');
+  ignore.addPatterns('!negated.txt');
+
+  // THEN
+  expect(splitAndIgnoreMarker(synthSnapshot(prj)['.myignorefile'])).toStrictEqual([
+    '*.bar',
+    '*.foo',
+    '!hello.foo',
+    '!negated.txt',
+    '!zoo.foo',
+  ]);
+});
+
+test('if include() is called with "!", then strip it', () => {
+  // GIVEN
+  const prj = new TestProject();
+  const ignore = new IgnoreFile(prj, '.myignorefile');
+
+  // WHEN
+  ignore.include('!*.js');
+
+  // THEN
+  expect(splitAndIgnoreMarker(synthSnapshot(prj)['.myignorefile'])).toStrictEqual([
+    '!*.js',
+  ]);
+});
+
+test('removePatters() can be used to remove previously added patters', () => {
+  // GIVEN
+  const prj = new TestProject();
+  const ignore = new IgnoreFile(prj, '.myignorefile');
+
+  // WHEN
+  ignore.addPatterns('*.js');
+  ignore.addPatterns('my_file');
+  ignore.addPatterns('!boom/bam');
+  ignore.removePatterns('*.zz', '*.js', '!boom/bam');
+  ignore.addPatterns('*.zz');
+  ignore.addPatterns('boom/bam');
+
+  // THEN
+  expect(splitAndIgnoreMarker(synthSnapshot(prj)['.myignorefile'])).toStrictEqual([
+    '*.zz',
+    'boom/bam',
+    'my_file',
+  ]);
+});
+
+test('comments are filtered out', () => {
+  // GIVEN
+  const prj = new TestProject();
+  const ignore = new IgnoreFile(prj, '.myignorefile');
+
+  // WHEN
+  ignore.addPatterns('*.js', '#comment');
+  ignore.addPatterns('!foo');
+  ignore.addPatterns('# hello world');
+  ignore.addPatterns('bar');
+
+  // THEN
+  expect(splitAndIgnoreMarker(synthSnapshot(prj)['.myignorefile'])).toStrictEqual([
+    '*.js',
+    'bar',
+    '!foo',
   ]);
 });