+++ src/__tests__/integration/cdk-watchful.projenrc.js
@@ -44,7 +44,7 @@ const project = new ConstructLibraryAws({
 
   // jsii publishing
 
-  java: {
+  publishToMaven: {
     javaPackage: 'com.github.eladb.watchful',
     mavenGroupId: 'com.github.eladb',
     mavenArtifactId: 'cdk-watchful',
@@ -52,7 +52,7 @@ const project = new ConstructLibraryAws({
     mavenRepositoryUrl: 'https://maven.pkg.github.com/eladb',
   },
 
-  python: {
+  publishToPypi: {
     distName: 'cdk-watchful',
     module: 'cdk_watchful',
   },