+++ src/index.ts
@@ -28,6 +28,7 @@ export * from './toml';
 export * from './typescript';
 export * from './version';
 export * from './yaml';
+export * from './xmlfile';
 
 // export submobules
 export * as tasks from './tasks';
@@ -35,3 +36,4 @@ export * as web from './web';
 export * as github from './github';
 export * as vscode from './vscode';
 export * as deps from './deps';
+export * as java from './java';